test-requirements.txt
setuptools
nose>=1.2.1
mock>=1.0.1
pyhamcrest>=1.6
